moved create config to tools
