+++ main.py
@@ -6,7 +6,7 @@ import re
 import praw
 import requests
 
-from tools import argument, blue, green, options, red, show_splash, verifyReddit
+from tools import argument, blue, green, options, red, show_splash, create_config
 
 VERSION = "0.1"
 
@@ -123,24 +123,6 @@ class R3dditScrapper:
         else:
             # restart the program if not using terminal arguments
             main()
-
-
-def create_config():
-    """Create config file if it doesn't exist"""
-    if not os.path.isfile("config.ini"):
-        config = configparser.ConfigParser()
-        config.add_section("Reddit")
-        client_id = input("Enter your client_id: ")
-        config.set("Reddit", "client_id", client_id)
-        client_secret = input("Enter your client_secret: ")
-        config.set("Reddit", "client_secret", client_secret)
-        with open("config.ini", "w") as f:
-            config.write(f)
-        if not verifyReddit(client_id, client_secret):
-            # credentials are invalid
-            print(red("Invalid credentials"))
-            os.remove("config.ini")
-            create_config()
 
 
 def setPath():
